new explosion sprite sheet

+++ DX11Base/billboard.h
@@ -10,7 +10,7 @@ private:
 	std::shared_ptr<BasicLightShader> m_shader;
 	ID3D11Buffer* m_vertexBuffer;
 	ID3D11ShaderResourceView* m_texture;
-	int m_count;
+	float m_count;
 
 public:
 	void Awake() override;
